requirements.txt
bs4==0.0.1
discord.py==1.5.0
python-dotenv==0.14.0